Move non-rest media files to a separate directory.
