+++ synapse/media/_base.py
