+++ synapse/media/storage_provider.py
