+++ synapse/media/thumbnailer.py
